fix: Resolve stdlib StringTools test failures with phi node and ABI fixes
Key fixes:
- Fix i32 to i64 extension for C ABI extern calls by checking Cranelift
  signature types instead of MIR types (cranelift_backend.rs)
- Register variable assignment results as locals for proper phi node
  tracking in while loops (hir_to_mir.rs)
- Implement proper do-while loop phi node handling for modified variables
- Add Array.join wrapper and runtime function (haxe_array_join)
- Add cca method mapping for StringTools.unsafeCodeAt support
- Update string runtime function signatures to use i64 for ABI compliance
- Remove unmapped StringTools functions (trim, ltrim, rtrim, isSpace)
  that should compile from Haxe source

All 47 stdlib tests now pass (24 parse + 23 functional).

+++ runtime/src/haxe_sys.rs
@@ -279,7 +279,7 @@ pub extern "C" fn haxe_string_len(s: *const HaxeString) -> i32 {
 /// Get character at index - returns empty string if out of bounds
 /// Note: charAt returns String, not Int, per Haxe specification
 #[no_mangle]
-pub extern "C" fn haxe_string_char_at_ptr(s: *const HaxeString, index: i32) -> *mut HaxeString {
+pub extern "C" fn haxe_string_char_at_ptr(s: *const HaxeString, index: i64) -> *mut HaxeString {
     if s.is_null() {
         return Box::into_raw(Box::new(HaxeString { ptr: std::ptr::null_mut(), len: 0, cap: 0 }));
     }
@@ -303,7 +303,7 @@ pub extern "C" fn haxe_string_char_at_ptr(s: *const HaxeString, index: i32) -> *
 
 /// Get character code at index - returns -1 (represented as null Int) if out of bounds
 #[no_mangle]
-pub extern "C" fn haxe_string_char_code_at_ptr(s: *const HaxeString, index: i32) -> i32 {
+pub extern "C" fn haxe_string_char_code_at_ptr(s: *const HaxeString, index: i64) -> i64 {
     if s.is_null() {
         return -1; // null
     }
@@ -312,7 +312,7 @@ pub extern "C" fn haxe_string_char_code_at_ptr(s: *const HaxeString, index: i32)
         if index < 0 || (index as usize) >= s_ref.len || s_ref.ptr.is_null() {
             return -1; // null
         }
-        *s_ref.ptr.add(index as usize) as i32
+        *s_ref.ptr.add(index as usize) as i64
     }
 }
